Add files via upload
This was an interesting problem; I spent a few days on it, dwelling here and there, and it is not the most efficient in terms of space, yet, I've capture a decent understanding of it. I shall come back to it, and refine it; in times like these, I like to appreciate learning, and forget very easily any kind perceived merit.